compare: better handling of moving field
Current diff heuristic does not work well when field are moved
around. It can result in a big diff for a small change. It's not wrong
per-se (there are often several possible ways to represent a change in
a diff), but can make the output nearly useless. This patch tries to
minimize the diff size for any changeset.