ksymtab: opaque internal data, abstract interface
Finish ksymtab modularization.

Provide abstract interface, which operates on opaque structure
and convert client code (generate) to use it: ksymtab and kabi
symbols operations.

The ksymtab module is not pure module since it combines pure
symtab abstraction (which is reused by kabi symbols operations)
and ksymtab reading, but to avoid big delta keeping it now as it
was.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>