generate: in-memory implementation reusing obj_t abstraction
This is a major refactoring of generate code. The idea is to
generate obj_t trees based on recursive parsing of dwarf
information and then dump them to the filesystem.

Some implementation details:

- introduce record_db abstraction which uses hash as the backend
  in-memory storage with previous filenames as the keys;
- introduce record abstraction as the database entry. obj_t
  itself is not enough to represent the previous highest level
  file data. Every record is linked to its root obj_t;
- records are using reference counter since once a record is
  produced, it can be
-- added to the database, may be with a new versioned name (and
   it should take ownership which is dropped on the final
   database cleanup), or
-- merged to the existing record and then it should be freed
   right away.

The main working horse, print_die() returns the obj_t for the die
argument. Every its helper returns an object for the particular
type. As previously, it can create a new record. In that case,
the return object is a reffile to the newly created record. For
the toplevel we do not connect it anywere, so just free it.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>