generate: print_die: split into three logical parts
Refactor huge print_die() to leave only top level logic there.

This patch prepares the code to record-based database backend
instead of raw filesystem for in-memory processing.

Split out print_die() into:

- record_start() which starts new a record (now yet file,
  temporary file in this case) and puts header info there;
- record_close() which finishes the record (closes the file);
- record_db_add() which addes the record to the db. For the
  existing logic it renames the temp file to the named one with
  possible merging and version increasing;

The patch shuffles the functions a bit to keep print_die*
functions family in one place and record processing before the
group.

The functions:

- print_stack_cb,
- get_header_line
- get_first_line

are moved as is.

It renames merge_file to record_merge, otherewise keeps it the
same.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>