generate: separate contexts from generate_config_t
There are different types of shared information:

- global (directory configuration, kabi symbols to find ...);
- per-processed module/vmlinux file (exported symbols);
- per-compilation unit (stack to reach a type, set of processed types);

At the moment some of it is a part of global structure (and works
since we process files sequentialy, some is passed as explicit
arguments.

Let's separate them from the global config and group into
structure to pass as one argument to the dependent funtions
(print_die family).

It will allow to improve the functions signatures a bit (one
structure instead of several argument), clean up the list of
passed arguments (actually, the Dwarf* handle is not needed at
that level) and helps in case of parallel processing of files or
compilation units.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>