generate: do not include full file name into record key
Currently the record key (and the final filename) contains the
full path to the declaration file: if "struct a" is declared in
the file a/b/c.h, the key will be "a/b/c.h/struct--a.txt".

There is no big point in it, the description file anyway contains
"File" tag with the place and most of the types anyway unique,
but having the prefix makes all the string operations slower.

Remove the prefix directories, as the result all the files will
be put in the root of the output directory, but it speeds up
generation on my system from 30s to 25s.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>