generate: make stack and cu generation optional
Currently the data records contain information about full
declaration stack of described type. The information is
debug-only and not used for the actual abi checks.

The compilation unit (CU) information can also be considered
debug-only, because it is not in use for actual comparations and
basically can be different depending of files scanning order,
since the same type can be described in many compilation units.

Make the generation optional, create a command-line switch
(-g, --generate-extra-info) to force its generation.

For CU to keep the object parser the same just put a fixed string
"nottracked" instead of the actual CU.

This improves generation speed on my setup by 1 second.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>