generate: use die copy for child iterations in subdies
struct, enum and union parsers used the passed die pointer to
store the child (member) elements and iteration on them. But the
data, it points to, should be actually reused on the upper
levels. For some reason the problem did not fire up initially,
but when I changed processing order it did.

Use own copy, like array parser does for them as well.

Since the needed libdwarf functions, dwarf_hasattr() in
particular, require non-const pointer, it's not easy to make the
parser function accepting const to make a compiler-time check.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>