generate: process EXPORT_SYMBOLed external named subprogs and vars only
Change the die filtering algorithm, get_symbol_index():

1) add a shortcut, return immediatly if die has not name. We
should process on the top level only EXPORT_SYMBOLed symbols, so
it cannot be unnamed;

2) change the inline filtering. We do not care about inline
keywords as soon as the symbol gets exported (since the macro
makes it external). The same symbol can be both exported and be
inlined in the original file: http://stackoverflow.com/a/216546 ;

3) change tag filtering actually according to the comment: accept
only subprogramms and variables on the top level and skip
types. All the needed types will be created recursively starting
from the subprogram/variable symbols.

(is_inline() is not needed anymore)

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>