ksymtab: implement symbol marking
This is a preparation for generation stub records for assembly
functions.

There is already symbols marking used externally to ksymtab
modules: in the kabi symbol list, passed via -s switch,
generation code marks symbols, which it finds in the debug info
and then reports the non-found ones.

To implement stub generation it will be needed to mark also
module's exported symbols, so it makes sence to implement that
generic logic inside of ksymtab module.

The patch keeps the module interface almost unchanged (for_each
changed) yet and generate code the same (except the for_each
call) for easier review.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>