ksymtab: elf independent __ksymtab_string section parser
Part of refactoring to separate section data handling from ELF
handling, which is a part of rework to implement symtab lookup
for weak symbols aliases.

Split the print_section() code into parse_ksymtab_string(), which
parses the raw data buffer of the section, and keep the ELF
section lookup separatly.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>