ksymtab: move elf setup/close to own open()/close() functions
Part of refactoring to separate section data handling from ELF
handling, which is a part of rework to implement symtab lookup
for weak symbols aliases.

Moves libelf elf file openning and basic checks to a separate
function. It does not yet create own elf object to keep the
incremental patchset simpler.

Move closing code to own function for the sake of consistency.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>