hash: make it possible to use binary keys with length
The current hash table implementation requires keys to be string,
while the hash algorithm itself doesn't really have such a
limitation.

For the weak aliases work we need addr->symbol mapping and
conversion address to string looks as a bit of overhead here.

Add support for binary keys also by adding _bin suffix to the
affected functions. Keep the old interface with a bunch of inline
wrappers on top of the new _bin interface, it should not affect
performance.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>