ksymtab: implement aliases finding
Alias is a symbol in ELF, which has the same value, as other
symbols.

We are interested in situation, when kernel exports a weak symbol
which is an alias of some non-exported global symbol (so the
exported symbol can be actully provided by some module, which
declares that symbol as global).

The main routine, ksymtab_read() will return now ksymtab of
aliases, the pointer is passed as the second argument. For the
patch generate.c doesn't use it yet, so passes just NULL.

There are two iterations: over ELF symbols and over ksymtab of
weak symbols, created by the first iteration, so there are 2
iteration callbacks: weak_filter() for ELF, and weak_to_alias()
for ksymtab.

First we iterate over all (well, WEAK and GLOBAL) symbols of
ELF. During the iteration two things are created (in one callback
to do it in one pass):
- all the global symbols form address -> global mapping.
  If there are several global symbols for the same address,
  the last found will be used for the final mapping;
- those weak symbols, which are exported, saved for further
  processing with their addresses;

Second time we are iterating over the weak symbols, saved during
the first iteration and using the map, tryig to find global
symbol for that address. If it exists, we create a symbol (named
as global) to return as alias (to the weak) and set its "link"
field to the name of the weak (exported) symbol.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>