parser: support special formats for assembly and weak files
Assembly records abused existing format which does not fit very
well to the needs. Also some format for weak aliases is required.

The patch adds 2 object types for that, output, compare, and
parser. Since there is no recursion in the records, the output is
simple.

The formats are:

Assembly: assembly <name>
Weak:     weak <weak_name> -> <global_name>

For parser it's better to keep the name in the data instead of
taking it from the filename.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>