ksymtab: export ksymtab_ksym_set_link()
The function will be used by the generator to save the alias.

Alias searching code returns global symbols pointing to the weak
aliases. But since the whitelisted symbols is weak, to generate
the weak record we need the back pointer. The whitelisted set is
constructed by the generation code, so we should look up the
alias there.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>