Introduce generate --rhel option to allow running on the RHEL build tree.
When kabi-dw is run on a RHEL kernel tree directory it might accidentaly read
same stale .ko modules from the redhat/rpm subdirectory. We need to ignore
that, so we introduce --rhel option which allows running on the RHEL kernel
tree.

We also change the -r/--replace-path to -a/--abs-path so the -r can be used for
--rhel.