generate: check specification die for external attribute
The compiler can generate two DIEs for a symbol: one for
declaration (it has DW_AT_declaration attribute) and another one
for the instance. The library combines them in the instance DIE
actually, but not fully: the DW_AT_external attribute is present
only in the declaration DIE. There is also similar configuration
for DW_AT_decl_file and DW_AT_decl_line for some symbols.

To not to skip symbols (we skip non-exported and declarations)
check the corresponding declaration DIE if it exists.

Fetch origin file information from the specification, if it is
missing from the main DIE and there is the specification.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>