Implement DW_AT_alignment support
The attribute can be present in different places. For example, it
can describe alignment of the whole struct, but also may present
in the individual member's DIEs.

Add to the record output additional field, Alignment -- this is
the top level one (for the whole structrue). Also change the
field format -- add the field alignment after its offset.

The alignment fields are added only if they are present in the DIE,
so not the default 0.

The patch also changes the parser to support the format in the
readers and reports the change in comparation code.

Signed-off-by: Yauheni Kaliuta <yauheni.kaliuta@redhat.com>