parser: don't use non C reserved word as keyword in the lexer
commit 3d2993f34b71 ("parser: support special formats for assembly and
weak files") added "assembly" and "weak" keyword to the lexer. However
neither are reserved word in C language and could be use as the name
of a variable, struct, etc. The parser would fails with a syntax error
if it ever encounter such symbol.

This two word are valid identifier, and recognised as such by the
lexer. Treat them as such and check that their value is the one
expected (as is currently done for the "CU" keyword). Also, factor
some redundant code.

Signed-off-by: Jerome Marchand <jmarchan@redhat.com>