parser: ignore unknown fields in the header
This reorganize the parser and lexer to take advantage of the new file
format. The code is now cleaner and better organized. From a
functional point of view, the only difference should be that the
parser now ignores unknown fields in the Please.

The lexer now uses start conditions to differentiates the header from
the symbol declaration. There is also a start condition for unknown
header field which are handled at the lexer level.

The parser now treats all files in the same way. Weak and assembly
symbol are handled in the symbol declaration section, as any other
type of symbol.

Signed-off-by: Jerome Marchand <jmarchan@redhat.com>