list: implement list structure
This commit introduces doubly linked list, with nodes having pointer to
the list containing them.

The nodes contain pointer to the list, they are contained in, so that we
don't have to store it separately when we want to remove the node from
the list. The pointer is needed in case the node is either the first or
the last node of the list, so that the corresponding members of the list
structure can be changed to the successor.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>