record, objects: link records and objects reffering to it
In the future a record, that is being referenced by an object, might get
merged and therefore destroyed, and if that happens all objects
referencing that object would be invalid, and so updating of references
to the record of those objects is needed. In order to update objects
referencing the now destroyed record, so that they are referencing the
record into which it was merged instead, there needs to be link between
the record and objects referencing it.

The link consists of a list of objects referencing the record on the
record side and a pointer to the node from previously mentioned list on
the object side.

Node from the list is required by the object, in case the object is
freed before dumping (e.g. if the object is merged), so that it can
remove the node from the list and avoid accessing of freed object.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>