generate: modify record_inc_version to record_set_version
Since there can be multiple types with the same name, we add to the
record name also its version, which is a number that distinguishes
records with otherwise same names and is appended after the name.

In preparation to storing multiple records under one list in hash map,
it would be faster to, after adding a new one, change their name just
once.

This will be possible because all the records, regardless of their
version, will be in one list and so that it is not required to create
new name string every time to find the next version of saved record.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>