generate: use list to store records
Multiple records are now kept in one list in record_db, this allows to
quickly find all similar records without having to create new string of
the record name every time.

Due to the planned merging of records at the end of program, where
records can be merged into a differently named record, the record to
which an object should reference may change. There needs to be a way to
change the references of all objects referencing to a certain record, so
that they are up-to-date.

New function record_update_dependents takes advantage of the link
between record and object, created by dependents field in record, to
update all objects referencing the record to the current name of the
record.

Since the way new records are added to the record_db changed there is no
longer need for record_db_lookup and record_db_push functions.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>