record, objects: manage declaration merging with parameter
Merging of objects that are only a declaration is undesirable before it
can be determined that all of the same name are same, that is anytime
before the final merge.

To achieve this the record_merge and subsequently obj_merge and
obj_merge_list functions now require additional argument merge_decl
specifying whether to merge objects that are declarations.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>