generate: wait with uncertain merging
The problem being solved is that multiple records can't be merged if it
is not possible to create groups in which they should be merged. This is
not a problem if all the records are able to be merged together, but
when there is one or more incompatible merging then the old way could
group them in the wrong way. And because the right way to group them
probably can't be decided, it is better not to group and subsequently
merge them at all.

This patch set aims to resolve this problem by waiting with undecidable
merging until the end, when it has all the versions of the record and
can decide if it can merge them into one or not. Meanwhile it merges
only those records that do not contain declarations are merged to reduce
the memory usage.

For merging all records in one list, there is a need to know if they can
really be merged beforehand, since otherwise they could end only
partially merged.

This can be accomplished by merging only those types that are completely
same and waiting with the rest after gathering all the records and being
clear if there is a possibility of them not being compatible.

Record_list_can_merge checks if every record on list could be merged
together by mock merging them with copy of one of the records. This is
important because we don't want to merge the list only partially.

Record_list_merge merges one list together into single record.

Record_db_merge uses the previous functions to merge every list
on record_db. We are merging until there are still list getting
merged, because there could be objects referencing another two records
that got merged after we would decide they are incompatible and so the
objects would stay unmerged.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>