generate: rework is_builtin() helper
The is_builtin() helper currently relies on a list of entries that are
known to be built-in. This is not portable as the list may vary from
arch to arch and may change in the future.  Instead check that the DIE
contains a DW_AT_decl_file, and if it does, whether its name ends in
"<built-in>".

This is not not perfect either, since AFAIK the "<built-in>"
string is provided by gcc and is not part of the DWARF
standard. It's however much better than the current status and
fixes the issue on aarch64.

Signed-off-by: Jerome Marchand <jmarchan@redhat.com>