generate: allow non-standard object file names
When a file (instead of a directory) is given as the parameter, whether
the file name ends with ".ko" or "vmlinux" shouldn't matter, to give us
more flexibility.

It also removes "..." from the "Generating symbol defs" message to make
it less confusing when the path is a file rather than a folder.

Signed-off-by: Xiao Jia <xiaoj@google.com>