generate: Implement reading Dwarf_Attribute value
Implement reading Dwarf_Attribute values for unsigned and block data.
This requires, especially in block data case, implementation of
location expression getting and evaluating which is only partially
implemented; basically all encountered operations with DWARF 2 through
to 5 with RHEL7 as well as RHEL8; relevant to reading:
 - DW_AT_{bit,byte}_size,
 - DW_AT_data_member_location,
 - DW_AT_alignment,
 - DW_AT_{data_,}_bit_offset
attributes and these should be read w/o a problem. Further operations
that could be handy were implemented as well.

This allows for reading of (not only) the said attributes w/o regard
for storage type (be it block or not).

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>