objects: use record for referencing in reffiles
Use record structure for referencing inside reffile obj_ts instead of
char *. This enables walking through the tree without having to slowly
find the record corresponding to the char *, and removes the need to
copy strings every version update. Since record structure is only used
while generating, this member is also only used in that case. When not
generating string with location of the record is saved in base_type
member.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>