utils: introduced global_string
Functions used for accessing static hash map, used for saving strings.
The strings cannot be modified and are freed all together, at the end of
the run of the program.

 - global_string_get_{copy,move}: takes a string to find in the hash map
 and returns a string with the same content from the hash map. Move
 version takes care of freeing the input string, copy does not free the
 original.

Using a single instance for every string with the same content enables
their comparision by comparing only their pointer without the need to
compare every character of the string.

Memory consumption is also lowered since global strings get rid of
duplicates and the lifetime of strings save here should be until dumping
of records, which is shortly followed by freeing global string hash.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>