generate: add function to merge two records
While still loading new debug information, pair of record trees is
merged if every pair of records contained within it meets merging
conditions (conditions are the same as with older merging, but versions
are ignored (since we also compare the referenced record) and the
declarations need to be at the same places (in order not to give
potentially false information)).

Function record_merge_pair merges a pair of records if they contain
declarations on the same members, arguments, etc. This function is used
while adding new batch of records from cu to db.

Also introduced a helping functions {obj,record}_same_declarations that
compares the declarations in a record tree. This function decreases the
time it takes to finish merging function if two records are not
mergeable due to different declarations.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>