generate: clear record_list nodes
An optimization that moves records that are often unable to be merged
with into a separate list. Records that were moved are reintroduced into
the main list after all records are loaded. Decreases number of records
to be checked.

Also clears list nodes that used to contain records that are now merged
into another one.

Signed-off-by: Štěpán Horáček <shoracek@redhat.com>