objects: add bytesize comparison to obj_eq
obj_eq does not perform a complete check of object equality. Consider:

 - structure packing

   two modules using a common header, exactly one built with -fpack-struct,
   or two modules built against an different versions of an external header
   file overtime, with exactly one explicitly using __attribute__((packed))

 - enum optimizations

   two modules using a common header, exactly one built with -fshort-enums

In cases such as these, bytesize entry will be emitted correctly for each
of these modules separately, however, when ran on both at the same time,
the missing obj_eq check causes the conflicting records to be merged.

Signed-off-by: Čestmír Kalina <ckalina@redhat.com>